+++ js/index.js
@@ -1,3 +1,3 @@
-import { init } from "./init.js";
+import { init } from "./eventListener/init.js";
 
 init();
